Calculating numbers with decimal points
These changes in the code will allow the calculator to take and perform calculations on expressions that have numbers with decimal point using Stack data structure.

+++ calcFrame.java
@@ -465,7 +465,7 @@ exp += addB.getText();
        if (result == null)
        screen.setText("Invalid Expression");                      
         else 
-        screen.setText(Integer.toString((int) result)); 
+        screen.setText(Double.toString((double) result)); 
 //exp = ""; // TODO add your handling code here:
     }
